Fix maskItUp
Rockstar has added a "Heist Accessories" line to the accessories submenu
whilst in heists, requiring an extra step to be taken for this macro.
I've updated the macro to now work in heists, but this will break it in
free roam causing it to swap helmets instead. The option is pretty much
useless in free roam anyway, though.